backend/app/requirements.txt
aiofiles==0.5.0
aiohttp
aiomysql==0.0.21
aioredis==1.3.1
alembic==1.4.2
async-timeout==3.0.1
celery==5.0.5
click==7.1.2
databases==0.4.3
emails==0.6
emoji==1.2.0
fastapi==0.66.0
fastapi-cache2==0.1.4
fastapi-utils==0.2.1
Jinja2==2.11.3
kombu==5.1.0
loguru==0.5.3
numpy==1.21.0
openpyxl==3.0.7
paramiko==2.7.2
passlib==1.7.4
psutil
pycryptodome==3.10.1
pyDes==2.0.1
PyExecJS==1.5.1
Pygments==2.9.0
PyJWT==2.1.0
pymongo==3.11.4
PyMySQL==0.9.3
pyOpenSSL==20.0.1
pyperclip==1.8.2
PySocks==1.7.1
python-editor==1.0.4
python-jose
python-multipart
python-socketio
PyYAML
redis==3.5.3
requests==2.23.0
rsa==4.7.2
six==1.14.0
SQLAlchemy==1.3.17
SQLAlchemy-Utils==0.37.8
starlette==0.14.2
urllib3==1.25.8
uvicorn==0.14.0
fastapi_profiler
motor
pydantic[email]
arq
rearq==0.2.4
tomlkit
arrow